requirements.txt
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
streamlit==1.32.0
python-dotenv==1.0.0
PyPDF2==3.0.1
psutil>=5.9.0